+++ server/src/main/java/com/linbit/linstor/core/apis/VolumeApi.java
@@ -2,6 +2,7 @@ package com.linbit.linstor.core.apis;
 
 import com.linbit.linstor.api.ApiCallRc;
 import com.linbit.linstor.api.interfaces.VlmLayerDataApi;
+import com.linbit.linstor.storage.kinds.DeviceLayerKind;
 import com.linbit.linstor.storage.kinds.DeviceProviderKind;
 import com.linbit.utils.Pair;
 
@@ -29,6 +30,13 @@ public interface VolumeApi
     @Deprecated
     /** returns the DeviceProviderKind of the storage pool of the vlmLayerObject with "" as resource name suffix */
     DeviceProviderKind getStorPoolDeviceProviderKind();
+
+    default Optional<StorPoolApi> getStorageStorPool()
+    {
+        return getVlmLayerData().stream()
+            .filter(entry -> entry.objA.equalsIgnoreCase(DeviceLayerKind.STORAGE.name())).findFirst()
+            .map(stringVlmLayerDataApiPair -> stringVlmLayerDataApiPair.objB.getStorPoolApi());
+    }
 
     ApiCallRc getReports();
 }